Move verify system up function to utils file
