+++ bubbles/commands/deploy.py
@@ -1,16 +1,14 @@
 import os
 import subprocess
-import time
 from typing import Callable
 
 from bubbles.commands import (
-    PROCESS_CHECK_COUNT,
-    PROCESS_CHECK_SLEEP_TIME,
     SERVICES,
     get_service_name,
 )
 from bubbles.config import PluginManager, COMMAND_PREFIXES
-from bubbles.utils import get_branch_head, say_code
+from bubbles.service_utils import verify_service_up, say_code
+from bubbles.utils import get_branch_head
 
 
 def _deploy_service(service: str, say: Callable) -> None:
@@ -54,22 +52,6 @@ def _deploy_service(service: str, say: Callable) -> None:
         say(f"Rolling back to previous state:\n```\n{git_response}```")
         subprocess.check_output(["sudo", "systemctl", "restart", get_service_name(loc)])
 
-    def verify_service_up(loc):
-        say(
-            f"Pausing for {PROCESS_CHECK_SLEEP_TIME}s to verify that {loc} restarted"
-            f" correctly..."
-        )
-        try:
-            for attempt in range(PROCESS_CHECK_COUNT):
-                time.sleep(PROCESS_CHECK_SLEEP_TIME / PROCESS_CHECK_COUNT)
-                subprocess.check_call(
-                    ["systemctl", "is-active", "--quiet", get_service_name(loc)]
-                )
-                say(f"Check {attempt + 1}/{PROCESS_CHECK_COUNT} complete!")
-            say("Restarted successfully!")
-        except subprocess.CalledProcessError:
-            revert_and_recover(loc)
-
     def restart_service(loc):
         say(f"Restarting service for {loc}...")
         systemctl_response = subprocess.check_output(
@@ -79,7 +61,10 @@ def _deploy_service(service: str, say: Callable) -> None:
             say("Something went wrong and could not restart.")
             say_code(say, systemctl_response)
         else:
-            verify_service_up(loc)
+            if verify_service_up(say, loc):
+                say("Restarted successfully!")
+            else:
+                revert_and_recover(loc)
 
     pull_from_git()
     try:
