+++ bubbles/commands/restart.py
@@ -1,38 +1,16 @@
 import subprocess
-import time
 from typing import Callable
 
 from bubbles.commands import (
-    PROCESS_CHECK_COUNT,
-    PROCESS_CHECK_SLEEP_TIME,
     SERVICES,
     get_service_name,
 )
 from bubbles.config import PluginManager, COMMAND_PREFIXES
-from bubbles.utils import say_code
+from bubbles.service_utils import say_code, verify_service_up
 
 
 def _restart_service(service: str, say: Callable) -> None:
     say(f"Restarting {service} in production. This may take a moment...")
-
-    def verify_service_up(loc):
-        say(
-            f"Pausing for {PROCESS_CHECK_SLEEP_TIME}s to verify that {loc} restarted"
-            f" correctly..."
-        )
-        try:
-            for attempt in range(PROCESS_CHECK_COUNT):
-                time.sleep(PROCESS_CHECK_SLEEP_TIME / PROCESS_CHECK_COUNT)
-                subprocess.check_call(
-                    ["systemctl", "is-active", "--quiet", get_service_name(loc)]
-                )
-                say(f"Check {attempt + 1}/{PROCESS_CHECK_COUNT} complete!")
-            say("Restarted successfully!")
-        except subprocess.CalledProcessError:
-            say(
-                f"{loc} is not responding. Cannot recover from here -- please check the"
-                f" logs for more information."
-            )
 
     def restart_service(loc):
         say(f"Restarting service for {loc}...")
@@ -43,7 +21,13 @@ def _restart_service(service: str, say: Callable) -> None:
             say("Something went wrong and could not restart.")
             say_code(say, systemctl_response)
         else:
-            verify_service_up(loc)
+            if verify_service_up(say, loc):
+                say("Restarted successfully!")
+            else:
+                say(
+                    f"{loc} is not responding. Cannot recover from here -- please check the"
+                    f" logs for more information."
+                )
 
     restart_service(service)
